# Optimize ds_align_comment() ##disasm