--- a/libr/include/r_util/r_str.h
+++ b/libr/include/r_util/r_str.h
@@ -37,6 +37,7 @@ R_API int r_str_bounds(const char *str, int *h);
 R_API char *r_str_crop(const char *str, unsigned int x, unsigned int y, unsigned int x2, unsigned int y2);
 R_API bool r_str_range_in(const char *r, ut64 addr);
 R_API int r_str_len_utf8(const char *s);
+R_API int r_str_len_utf8_ansi(const char *str);
 R_API int r_str_len_utf8char(const char *s, int left);
 R_API void r_str_filter_zeroline(char *str, int len);
 R_API int r_str_utf8_codepoint(const char *s, int left);