`ds_align_comment()` is currently the bottleneck for `agJ`. The idea here is to save the position in the cons buffer of the start of the current line and calculate the line length starting from there, excluding ansi escape sequences and respecting utf-8.

## Comparison
Before:
```
[florian@florian-desktop radare2]$ time r2 /bin/ls -qc "s main;af;agJ" > /dev/null

real	0m0,440s
user	0m0,428s
sys	0m0,012s
```

After:
```
[florian@florian-desktop radare2]$ time r2 /bin/ls -qc "s main;af;agJ" > /dev/null

real	0m0,153s
user	0m0,139s
sys	0m0,013s
```

Here is the comparison of a seek in cutter from outside a function to inside a function, which causes a call to `agJ`:
Before:
![bildschirmfoto vom 2018-12-25 12-50-25](https://user-images.githubusercontent.com/1460997/50421882-45998c80-0844-11e9-9840-4c12dbbc4179.png)

After:
![bildschirmfoto vom 2018-12-25 12-50-42](https://user-images.githubusercontent.com/1460997/50421885-4af6d700-0844-11e9-968a-65e28302ad9a.png)