@radare 2018-12-25
💥

> On 25 Dec 2018, at 14:27, Codecov <notifications@github.com> wrote:
> 
> Codecov Report
> 
> Merging #12560 into master will decrease coverage by <.01%.
> The diff coverage is 83.33%.
> 
> 
> 
> @@            Coverage Diff             @@
> ##           master   #12560      +/-   ##
> ==========================================
> - Coverage   36.99%   36.98%   -0.01%     
> ==========================================
>   Files         907      907              
>   Lines      290608   290622      +14     
> ==========================================
> - Hits       107509   107500       -9     
> - Misses     183099   183122      +23
> Impacted Files	Coverage Δ	
> libr/include/r_util/r_str.h	100% <ø> (ø)	⬆️
> libr/cons/cons.c	54.4% <100%> (-2.49%)	⬇️
> libr/core/disasm.c	72.74% <81.25%> (-0.03%)	⬇️
> libr/util/str.c	60.18% <91.66%> (+0.46%)	⬆️
> libr/debug/dreg.c	65.9% <0%> (-1.37%)	⬇️
> libr/core/cmd_anal.c	47.39% <0%> (-0.03%)	⬇️
> Continue to review full report at Codecov.
> 
> Legend - Click here to learn more
> Δ = absolute <relative> (impact), ø = not affected, ? = missing data
> Powered by Codecov. Last update e7451c6...27d72f9. Read the comment docs.
> 
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub, or mute the thread.