@radare 2018-12-25
Good optimization

> On 25 Dec 2018, at 23:27, Florian Märkl <notifications@github.com> wrote:
> 
> wat
> 
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub, or mute the thread.